ref: renamed middleware to be more specific
