+++ middleware/authenticateUser.ts
